pyproject.toml
[tool.poetry]
name = "longanisa-maker"
version = "0.1.0"
description = ""
authors = ["Emmanuel Rustia Pagayonan <emmanuel_pagayonan_28@yahoo.com>"]
readme = "README.md"
packages = [{include = "longanisa_maker"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
whitenoise = "^6.6.0"
dj-rest-auth = "^4.0.1"
drf-yasg = "^1.21.6"
djangorestframework-simplejwt = "^5.2.2"
django-allauth = "^0.54.0"
django-cors-headers = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
